Refactor presence roles into plugins
