+++ betty/tests/extension/cotton_candy/assets/templates/entity/test_label_event_html_j2.py
@@ -1,6 +1,7 @@
 from betty.extension.cotton_candy import CottonCandy
 from betty.jinja2 import EntityContexts
-from betty.model.ancestry import Person, Event, Presence, Subject, Witness
+from betty.model.ancestry import Person, Event, Presence
+from betty.model.presence_role import Subject, Witness
 from betty.model.event_type import Birth, Marriage
 from betty.tests import TemplateTestCase
