+++ betty/tests/extension/deriver/test___init__.py
@@ -5,7 +5,8 @@ from betty.load import load
 from betty.locale import DateRange, Date
 from betty.locale.localizable import Localizable, plain
 from betty.model import record_added
-from betty.model.ancestry import Person, Presence, Subject, Event
+from betty.model.ancestry import Person, Presence, Event
+from betty.model.presence_role import Subject
 from betty.model.event_type import (
     DerivableEventType,
     CreatableDerivableEventType,
